Fix detail blog and portfolio

+++ src/app/routes/admin/subscribes/subscribes.component.ts
@@ -1,9 +1,11 @@
-import { Component, Inject } from "@angular/core";
+import { Component, Inject, ViewChild } from "@angular/core";
 import { NgbModal, ModalDismissReasons } from "@ng-bootstrap/ng-bootstrap";
 import { ConfirmModalComponent } from "@components/admin/confirm-modal/confirm-modal.component";
 import { TagService } from "app/services/tag.service";
 import { Subject } from "rxjs";
 import { SubscribeService } from "@services/subscribe.service";
+import { DataTableDirective } from "angular-datatables";
+
 @Component({
 	selector: "app-admin-subscribes",
 	templateUrl: "./subscribes.component.html",
@@ -16,6 +18,8 @@ export class SubscribesComponent {
 	dtOptions = {};
 	dtTrigger = new Subject();
 	sub: any;
+	index = 0;
+	@ViewChild(DataTableDirective) dtElement: DataTableDirective;
 
 	constructor(
 		private modalService: NgbModal,
@@ -28,16 +32,32 @@ export class SubscribesComponent {
 
 	getAll() {
 		this.sub = this.subscribeService.getAll().subscribe(subscribes => {
-			//console.log(subscribes);
 			this.subscribes = subscribes;
+			if (this.index == 0) {
+				this.dtTrigger.next();
+				this.index++;
+			} else {
+				this.rerender();
+			}
 		});
 	}
-	onDelete(subscribe){
-		this.subscribeService.delete(subscribe.id);
+	onDelete(subscribe) {
+		const modalRef = this.modalService.open(ConfirmModalComponent);
+		modalRef.componentInstance.title = "Subscribe delete";
+		modalRef.result.then(result => {
+			if (result === "ok") {
+				this.subscribeService.delete(subscribe.id);
+			}
+		});
 	}
-	
 
 	ngOnDestroy(): void {
 		this.sub.unsubscribe();
 	}
+	rerender = () => {
+		this.dtElement.dtInstance.then((dtInstance: DataTables.Api) => {
+			dtInstance.destroy();
+			this.dtTrigger.next();
+		});
+	};
 }
